Fix issue #2 (multiple monitors)
Originally, i3lock-next-helper assumes that monitors are in order and
calculates the position of the next monitor using the width and height
of the previous monitor. If we instead grab the monitor position data,
this assumption doesn't need to be made and we can just use the (x,y)
position data for each monitor.

So now, for each monitor, we grab the resolution (widths[] and
heights[]) as well as the position (xcoords[] and ycoords[]) using
liXrandr's XRRGetCrtcInfo(). With this, we can calculate the centre of
each monitor. (Because, from xcoords[] and ycoords[], we know the
coordinates of the top left corner of the monitor.)

This commit will also move the code for the case of a single monitor
into the loop originally for multiple monitors as nothing different
needs to be done.